Add gcab_cabinet_decompress()
This gives GCab to ability to decompress each file to a memory blob.

At the moment fwupd decompresses each archive into a folder in /tmp, then reads
back the files into memory so that it can parse them. This is fragile, and
somewhat insecure if not done carefully.

Allowing GCab to decompress to a GBytes buffer means we can make the fwupd code
a lot simpler and safer.